Add Leaflet.draw polygon filter
This filter was broken due to two issues.

1. The missing css caused the 'Add vertex' marker
added by the polyline class to be unclickable, which
breaks the logic that saves the click point in the
onmousedown event:
https://github.com/Leaflet/Leaflet.draw/blob/master/src/draw/handler/Draw.Polyline.js#L253

2. The default zIndexOffset wasn't high enough to place
the draw layer on top of all others, masking the mouse
events.

Also updates the polygon draw options to more sensible defaults.